pyproject.toml
[project]
name = "cartratbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Egor Khudiakov",email = "eh83865@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytelegrambotapi (>=4.26.0,<5.0.0)",
    "django (>=5.2,<6.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
packages = [{include = "cartratbot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
